Moves interceptors to the "interceptors" directory
